# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Raasta Pakistan - Live Traffic & Offers Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /offers:
    get:
      operationId: listOffers
      tags: [offers]
      summary: List active offers, optionally near a location
      parameters:
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: category
          in: query
          schema:
            type: string
        - name: merchantId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
    post:
      operationId: createOffer
      tags: [offers]
      summary: Create a new merchant offer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOfferBody"
      responses:
        "201":
          description: Offer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /offers/{id}:
    get:
      operationId: getOffer
      tags: [offers]
      summary: Get a specific offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Offer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
    patch:
      operationId: updateOffer
      tags: [offers]
      summary: Update an offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOfferBody"
      responses:
        "200":
          description: Updated offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
    delete:
      operationId: deleteOffer
      tags: [offers]
      summary: Delete an offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /offers/{id}/redeem:
    post:
      operationId: redeemOffer
      tags: [offers]
      summary: Redeem an offer (generate QR code token)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedeemOfferBody"
      responses:
        "200":
          description: Redemption token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedemptionToken"
...
components:
  schemas:
...
    Offer:
      type: object
      properties:
        id:
          type: string
        merchantId:
          type: string
        merchantName:
          type: string
        merchantCategory:
          type: string
        title:
          type: string
        description:
          type: string
        discountPercent:
          type: number
        discountAmount:
          type: number
        originalPrice:
          type: number
        offerPrice:
          type: number
        code:
          type: string
        imageUrl:
          type: string
        lat:
          type: number
        lng:
          type: number
        address:
          type: string
        city:
          type: string
        radiusKm:
          type: number
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
        maxRedemptions:
          type: integer
        currentRedemptions:
          type: integer
        isActive:
          type: boolean
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - merchantId
        - merchantName
        - merchantCategory
        - title
        - description
        - lat
        - lng
        - isActive
        - createdAt
    CreateOfferBody:
      type: object
      properties:
        merchantId:
          type: string
        title:
          type: string
        description:
          type: string
        discountPercent:
          type: number
        discountAmount:
          type: number
        originalPrice:
          type: number
        offerPrice:
          type: number
        code:
          type: string
        imageUrl:
          type: string
        radiusKm:
          type: number
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
        maxRedemptions:
          type: integer
        tags:
          type: array
          items:
            type: string
      required: [merchantId, title, description]
    UpdateOfferBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        discountPercent:
          type: number
        discountAmount:
          type: number
        isActive:
          type: boolean
        validUntil:
          type: string
          format: date-time
        maxRedemptions:
          type: integer
    RedeemOfferBody:
      type: object
      properties:
        userId:
          type: string
        userName:
          type: string
      required: [userId]
    RedemptionToken:
      type: object
      properties:
        token:
          type: string
        offerId:
          type: string
        expiresAt:
          type: string
          format: date-time
        qrData:
          type: string
      required: [token, offerId, expiresAt, qrData]
...