# lib/api-zod/src/generated/types/redeemOfferBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Raasta Pakistan - Live Traffic & Offers Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RedeemOfferBody {
  userId: string;
  userName?: string;
}