# lib/api-zod/src/generated/types/updateOfferBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Raasta Pakistan - Live Traffic & Offers Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateOfferBody {
  title?: string;
  description?: string;
  discountPercent?: number;
  discountAmount?: number;
  isActive?: boolean;
  validUntil?: Date;
  maxRedemptions?: number;
}